backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
groq
passlib
python-jose[cryptography]
python-multipart
